requirements.txt
boto3==1.11.9
botocore==1.14.9
Cerberus==1.3.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
dask==2.3.0
decorator==4.4.1
distributed==2.1.0
docutils==0.15.2
etelemetry==0.1.2
filelock==3.0.12
fsspec==0.6.2
funcsigs==1.0.2
future==0.18.2
gitdb2==2.0.6
GitPython==3.0.5
HeapDict==1.0.1
idna==2.8
isodate==0.6.0
jmespath==0.9.4
lxml==4.5.0
msgpack==0.6.2
networkx==2.3
neurdflib==5.0.1
nibabel==3.0.1
nipype==1.2.3
numpy==1.18.1
packaging==20.1
prov==1.5.3
psutil==5.6.7
pydot==1.4.1
pydotplus==2.0.2
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3
rdflib==4.2.2
requests==2.22.0
s3fs==0.4.0
s3transfer==0.3.2
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
smmap2==2.0.5
sortedcontainers==2.1.0
tblib==1.6.0
toolz==0.10.0
tornado==6.0.3
traits==5.2.0
urllib3==1.25.8
xvfbwrapper=0.2.9
zict==1.0.0
